+++ Assets/TurnrootFramework/CharacterClass/Components/Validation.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 233ef7aecb8d30c468c8fcd985a316be
