+++ Assets/TurnrootFramework/Gameplay/Combat/FundamentalComponents/Context/Components.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2472284123b2d5e45b10bcd3d18b25e6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
